a2a-adk-app/pyproject.toml
[project]
name = "a2a-mcp-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.2",
    "click>=8.2.0",
    "geopy>=2.4.1",
    "gradio>=5.30.0",
    "langchain-mcp-adapters>=0.1.1",
    "langgraph>=0.4.5",
    "mcp>=1.9.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-core>=0.1.23",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "huggingface-hub>=0.20.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "accelerate>=0.25.0",
    "bitsandbytes>=0.41.0",
    "transformers>=4.30.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
]
